# docs/swagger.yaml
openapi: 3.0.2
info:
    title: Games-api
    version: 1.0.0
    description: '### This is a Swagger UI for our Games API'
paths:
    /games-api:
        description: ''
        get:
            tags:
                - games
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of games:
                                    value:
                                        - Witcher 3
                                        - Cyberpunk2077
                    description: List of games
            summary: Get a list of all games
            description: Returns a list of all games in API
        post:
            requestBody:
                description: A new game to be created
                required: true
            tags:
                - new-game
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/new-game'
                    description: Create new game
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                    description: One or all params are missing
            summary: Create a game
            description: Create a new game
    '/games/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Full details of a given game:
                                    value:
                                        id: 1
                                        name: Witcher 3
                                        price: 29.99
                    description: Returns full details of a given id game
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Game /w + was not found:
                                    value:
                                        error: Game\w+ was not found in game array.
                    description: Game was not found in games array
            operationId: id
            summary: Get full details of game
            description: Returns full details of a given id game
        delete:
            requestBody: {}
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/delete-game'
                    description: All good
                '204':
                    description: Content not found
                '404':
                    content:
                        application/json:
                            examples:
                                Example_error:
                                    value:
                                        error: Cannot find game
                    description: Not found
            summary: Delete game
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.
        new-game:
            title: Root Type for new-game
            description: ''
            required:
                - name
                - price
            type: object
            properties:
                name:
                    type: string
                price:
                    format: double
                    type: number
            example:
                name: Tetris
                price: 0.99
        delete-game:
            title: Root Type for delete-game
            description: ''
            required:
                - id
            type: object
            properties:
                id:
                    format: int32
                    type: integer
            example:
                id: 1